requirements.txt
jax>=0.4.0
jaxlib>=0.4.0
flax>=0.6.0
optax>=0.1.4
pyscf>=2.3.0
numpy>=1.24.0
torch>=2.0.0
tensorboard>=2.13.0
orbax-checkpoint>=0.4.0
tqdm>=4.65.0
h5py>=3.8.0
pandas>=2.0.0
openpyxl>=3.1.0
